Fixed scripts. tested repos and repos-test now
git-svn-id: https://optime.repos.se/svn/code/repos/trunk@3123 94f62164-8fde-0310-83d7-e1105c8c1596

+++ build/debian-packages/repos-test.arm/var/www/admin/switch-to-data.sh
@@ -11,9 +11,16 @@ then
  rm -v /etc/apache2/sites-enabled/testrepo.conf
 fi
 
-echo "Switch repos settings"
-mv -v /var/www/admin/repos.properties /var/www/admin/repos.properties.test
-mv -v /var/www/admin/repos.properties.data /var/www/admin/repos.properties
-mv -v /var/www/admin/repos-access.data /var/www/admin/repos-access
+if [ -f /var/www/admin/repos.properties.data ]
+then
+ echo "Switch repos settings"
+ mv -v /var/www/admin/repos.properties /var/www/admin/repos.properties.test
+ mv -v /var/www/admin/repos.properties.data /var/www/admin/repos.properties
+fi
+
+if [ -f /var/www/admin/repos-access.data ]
+then
+ mv -v /var/www/admin/repos-access.data /var/www/admin/repos-access
+fi 
 
 /etc/init.d/apache2 reload
